Split out common pfc test functions to common/
